Restructure locust test
